+++ qiskit_optimization/algorithms/goemans_williamson_optimizer.py
@@ -143,9 +143,8 @@ class GoemansWilliamsonOptimizer(OptimizationAlgorithm):
 
         cuts = self._generate_random_cuts(chi, len(adj_matrix))
 
-        from qiskit_optimization.applications.max_cut import Maxcut
         numeric_solutions = [(cuts[i, :],
-                              Maxcut.max_cut_value(cuts[i, :], adj_matrix))
+                              self.max_cut_value(cuts[i, :], adj_matrix))
                              for i in range(self._num_cuts)]
 
         if self._sort_cuts:
@@ -261,3 +260,17 @@ class GoemansWilliamsonOptimizer(OptimizationAlgorithm):
         r = np.random.normal(size=(self._num_cuts, num_vertices))
 
         return (np.dot(r, x) > 0) + 0
+
+    @staticmethod
+    def max_cut_value(x: np.ndarray, adj_matrix: np.ndarray):
+        """Compute the value of a cut from an adjacency matrix and a list of binary values.
+
+        Args:
+            x: a list of binary value in numpy array.
+            adj_matrix: adjacency matrix.
+
+        Returns:
+            float: value of the cut.
+        """
+        cut_matrix = np.outer(x, (1 - x))
+        return np.sum(adj_matrix * cut_matrix)
