+++ test/applications/test_knapsack.py
@@ -65,6 +65,7 @@ class TestKnapsack(QiskitOptimizationTestCase):
         self.assertEqual(knapsack.interpret(self.result), [1, 3])
 
     def test_max_weight(self):
+        """Test max_weight"""
         knapsack = Knapsack(values=self.values, weights=self.weights, max_weight=self.max_weight)
         knapsack.max_weight = 5
         self.assertEqual(knapsack.max_weight, 5)
